Update patch set 4
Patch Set 4:

So the real "workaround" for the venus buffer size change to not break playback is only disabling the fatal error in the size check.
The rest is just a better "rewrite" of https://gerrit.omnirom.org/#/c/10607/
Interesstingly the diff there is always 4k but venus extrabuffer size hack changed from 8k to 0 :) Furthermore the fatal error in that size change actually looks not-fatal since if the buffer is smaller then expected and the value is adjusted afterwards.

Patch-set: 4